+++ swing/framework-model-tools/src/test/java/is/codion/swing/framework/model/tools/generator/DomainGeneratorModelTest.java
@@ -20,7 +20,6 @@ package is.codion.swing.framework.model.tools.generator;
 
 import is.codion.common.db.database.Database;
 import is.codion.common.user.User;
-import is.codion.swing.framework.model.tools.metadata.MetaDataSchema;
 
 import org.junit.jupiter.api.Test;
 
@@ -29,6 +28,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.Comparator;
 
+import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.joining;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -40,17 +40,21 @@ public final class DomainGeneratorModelTest {
 	@Test
 	void test() throws Exception {
 		DomainGeneratorModel model = DomainGeneratorModel.domainGeneratorModel(Database.instance(), UNIT_TEST_USER);
-		model.domainPackage().set("is.codion.test.petstore.domain");
+		model.domainPackage().set("is.codion.petstore.domain");
 		model.schemaModel().refresh();
-		model.schemaModel().comparator().set(Comparator.comparing(MetaDataSchema::name));
+		model.schemaModel().comparator().set(comparing(SchemaRow::name));
 		model.schemaModel().selectionModel().setSelectedIndex(1);
 		model.populateSelected(schema -> {});
 		model.definitionModel().comparator().set(Comparator.
 						<DefinitionRow, String>comparing(definitionRow -> definitionRow.definition.entityType().domainType().name())
 						.thenComparing(definitionRow -> definitionRow.definition.entityType().name()));
 		model.definitionModel().selectionModel().selectAll();
-		String petstore = textFileContents(DomainGeneratorModelTest.class, "Petstore.java");
-		assertEquals(petstore, model.domainSource().get());
+		String petstoreApi = textFileContents(DomainGeneratorModelTest.class, "PetstoreAPI.java");
+		assertEquals(petstoreApi, model.domainApi().get());
+		String petstoreImpl = textFileContents(DomainGeneratorModelTest.class, "PetstoreImpl.java");
+		assertEquals(petstoreImpl, model.domainImpl().get());
+		String petstoreCombined = textFileContents(DomainGeneratorModelTest.class, "Petstore.java");
+		assertEquals(petstoreCombined, model.domainCombined().get());
 		model.close();
 	}