+++ swing/framework-model-tools/src/test/resources/is/codion/swing/framework/model/tools/generator/PetstoreAPI.java
@@ -1,0 +1,102 @@
+package is.codion.petstore.domain;
+
+import static is.codion.framework.domain.DomainType.domainType;
+
+import is.codion.framework.domain.DomainType;
+import is.codion.framework.domain.entity.EntityType;
+import is.codion.framework.domain.entity.attribute.Column;
+import is.codion.framework.domain.entity.attribute.ForeignKey;
+
+public interface Petstore {
+	DomainType DOMAIN = domainType(Petstore.class);
+
+	interface Address {
+		EntityType TYPE = DOMAIN.entityType("petstore.address");
+
+		Column<Integer> ADDRESS_ID = TYPE.integerColumn("address_id");
+		Column<String> STREET1 = TYPE.stringColumn("street1");
+		Column<String> STREET2 = TYPE.stringColumn("street2");
+		Column<String> CITY = TYPE.stringColumn("city");
+		Column<String> STATE = TYPE.stringColumn("state");
+		Column<Integer> ZIP = TYPE.integerColumn("zip");
+		Column<Double> LATITUDE = TYPE.doubleColumn("latitude");
+		Column<Double> LONGITUDE = TYPE.doubleColumn("longitude");
+		Column<Object> LOCATION = TYPE.column("location", Object.class);
+		Column<byte[]> IMAGE = TYPE.byteArrayColumn("image");
+	}
+
+	interface Category {
+		EntityType TYPE = DOMAIN.entityType("petstore.category");
+
+		Column<Integer> CATEGORY_ID = TYPE.integerColumn("category_id");
+		Column<String> NAME = TYPE.stringColumn("name");
+		Column<String> DESCRIPTION = TYPE.stringColumn("description");
+		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+	}
+
+	interface ContactInfo {
+		EntityType TYPE = DOMAIN.entityType("petstore.contact_info");
+
+		Column<Integer> CONTACT_INFO_ID = TYPE.integerColumn("contact_info_id");
+		Column<String> LAST_NAME = TYPE.stringColumn("last_name");
+		Column<String> FIRST_NAME = TYPE.stringColumn("first_name");
+		Column<String> EMAIL = TYPE.stringColumn("email");
+	}
+
+	interface Item {
+		EntityType TYPE = DOMAIN.entityType("petstore.item");
+
+		Column<Integer> ITEM_ID = TYPE.integerColumn("item_id");
+		Column<Integer> PRODUCT_ID = TYPE.integerColumn("product_id");
+		Column<String> NAME = TYPE.stringColumn("name");
+		Column<String> DESCRIPTION = TYPE.stringColumn("description");
+		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+		Column<String> IMAGE_THUMB_URL = TYPE.stringColumn("image_thumb_url");
+		Column<Double> PRICE = TYPE.doubleColumn("price");
+		Column<Integer> ADDRESS_ID = TYPE.integerColumn("address_id");
+		Column<Integer> CONTACT_INFO_ID = TYPE.integerColumn("contact_info_id");
+		Column<Integer> TOTAL_SCORE = TYPE.integerColumn("total_score");
+		Column<Integer> NUMBER_OF_VOTES = TYPE.integerColumn("number_of_votes");
+		Column<Integer> DISABLED = TYPE.integerColumn("disabled");
+
+		ForeignKey PRODUCT_ID_FK = TYPE.foreignKey("product_id_fk", PRODUCT_ID, Product.PRODUCT_ID);
+		ForeignKey ADDRESS_ID_FK = TYPE.foreignKey("address_id_fk", ADDRESS_ID, Address.ADDRESS_ID);
+		ForeignKey CONTACT_INFO_ID_FK = TYPE.foreignKey("contact_info_id_fk", CONTACT_INFO_ID, ContactInfo.CONTACT_INFO_ID);
+	}
+
+	interface ItemTags {
+		EntityType TYPE = DOMAIN.entityType("petstore.item_tags");
+
+		Column<String> NAME = TYPE.stringColumn("name");
+		Column<String> TAG = TYPE.stringColumn("tag");
+	}
+
+	interface Product {
+		EntityType TYPE = DOMAIN.entityType("petstore.product");
+
+		Column<Integer> PRODUCT_ID = TYPE.integerColumn("product_id");
+		Column<Integer> CATEGORYID = TYPE.integerColumn("categoryid");
+		Column<String> NAME = TYPE.stringColumn("name");
+		Column<String> DESCRIPTION = TYPE.stringColumn("description");
+		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+
+		ForeignKey CATEGORYID_FK = TYPE.foreignKey("categoryid_fk", CATEGORYID, Category.CATEGORY_ID);
+	}
+
+	interface Tag {
+		EntityType TYPE = DOMAIN.entityType("petstore.tag");
+
+		Column<Integer> TAG_ID = TYPE.integerColumn("tag_id");
+		Column<String> TAG = TYPE.stringColumn("tag");
+	}
+
+	interface TagItem {
+		EntityType TYPE = DOMAIN.entityType("petstore.tag_item");
+
+		Column<Integer> TAG_ID = TYPE.integerColumn("tag_id");
+		Column<Integer> ITEM_ID = TYPE.integerColumn("item_id");
+
+		ForeignKey TAG_ID_FK = TYPE.foreignKey("tag_id_fk", TAG_ID, Tag.TAG_ID);
+		ForeignKey ITEM_ID_FK = TYPE.foreignKey("item_id_fk", ITEM_ID, Item.ITEM_ID);
+	}
+}