+++ swing/framework-ui-tools/src/main/java/is/codion/swing/framework/ui/tools/generator/DomainGeneratorPanel.java
@@ -19,6 +19,7 @@
 package is.codion.swing.framework.ui.tools.generator;
 
 import is.codion.common.db.database.Database;
+import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.i18n.Messages;
 import is.codion.common.model.CancelException;
 import is.codion.swing.common.ui.Utilities;
@@ -35,7 +36,7 @@ import is.codion.swing.framework.model.tools.generator.DefinitionRow;
 import is.codion.swing.framework.model.tools.generator.DomainGeneratorModel;
 import is.codion.swing.framework.model.tools.generator.DomainGeneratorModel.DefinitionColumns;
 import is.codion.swing.framework.model.tools.generator.DomainGeneratorModel.SchemaColumns;
-import is.codion.swing.framework.model.tools.metadata.MetaDataSchema;
+import is.codion.swing.framework.model.tools.generator.SchemaRow;
 
 import com.formdev.flatlaf.intellijthemes.FlatAllIJThemes;
 
@@ -49,6 +50,7 @@ import javax.swing.JTextArea;
 import javax.swing.SortOrder;
 import javax.swing.SwingConstants;
 import javax.swing.SwingUtilities;
+import javax.swing.UIManager;
 import javax.swing.WindowConstants;
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
@@ -84,7 +86,7 @@ public final class DomainGeneratorPanel extends JPanel {
 						.name("Populate")
 						.enabled(model.schemaModel().selectionModel().selectionNotEmpty())
 						.build();
-		FilterTable<MetaDataSchema, SchemaColumns.Id> schemaTable =
+		FilterTable<SchemaRow, SchemaColumns.Id> schemaTable =
 						FilterTable.builder(model.schemaModel(), createSchemaColumns())
 										.autoResizeMode(JTable.AUTO_RESIZE_ALL_COLUMNS)
 										.doubleClickAction(populateSchemaControl)
@@ -114,11 +116,14 @@ public final class DomainGeneratorPanel extends JPanel {
 						.bottomComponent(new JScrollPane(definitionTable))
 						.build();
 
+		JLabel packageLabel = label("Package")
+						.displayedMnemonic('P')
+						.build();
 		JPanel packagePanel = borderLayoutPanel()
-						.westComponent(label("Package").build())
+						.westComponent(packageLabel)
 						.centerComponent(stringField(model.domainPackage())
 										.columns(42)
-										.build())
+										.build(packageLabel::setLabelFor))
 						.build();
 
 		JPanel schemaTablePanel = borderLayoutPanel()
@@ -128,38 +133,87 @@ public final class DomainGeneratorPanel extends JPanel {
 						.centerComponent(schemaTableSplitPane)
 						.build();
 
-		JTextArea sourceTextArea = textArea()
+		Font font = UIManager.getFont("TextArea.font");
+		Font monospace = new Font(Font.MONOSPACED, font.getStyle(), font.getSize());
+		JTextArea apiTextArea = textArea()
 						.rowsColumns(40, 60)
 						.editable(false)
+						.font(monospace)
+						.build();
+		JTextArea implementationTextArea = textArea()
+						.rowsColumns(40, 60)
+						.editable(false)
+						.font(monospace)
+						.build();
+		JTextArea combinedTextArea = textArea()
+						.rowsColumns(40, 60)
+						.editable(false)
+						.font(monospace)
 						.build();
 
-		Font font = sourceTextArea.getFont();
-		sourceTextArea.setFont(new Font(Font.MONOSPACED, font.getStyle(), font.getSize()));
-		JPanel textAreaPanel = borderLayoutPanel()
-						.northComponent(borderLayoutPanel()
-										.eastComponent(packagePanel)
-										.build())
-						.centerComponent(new JScrollPane(sourceTextArea))
-						.southComponent(flowLayoutPanel(FlowLayout.RIGHT)
-										.add(button(Control.builder(() -> Utilities.setClipboard(sourceTextArea.getText()))
-														.name(Messages.copy()))
-														.mnemonic(Messages.copy().charAt(0))
+		JSplitPane apiImplPanel = splitPane()
+						.orientation(JSplitPane.VERTICAL_SPLIT)
+						.resizeWeight(0.5)
+						.topComponent(borderLayoutPanel()
+										.centerComponent(new JScrollPane(apiTextArea))
+										.southComponent(flowLayoutPanel(FlowLayout.RIGHT)
+														.add(button(createCopyControl(apiTextArea))
+																		.build())
 														.build())
 										.build())
+						.bottomComponent(borderLayoutPanel()
+										.centerComponent(new JScrollPane(implementationTextArea))
+										.southComponent(flowLayoutPanel(FlowLayout.RIGHT)
+														.add(button(createCopyControl(implementationTextArea))
+																		.build())
+														.build())
+										.build())
+						.continuousLayout(true)
+						.oneTouchExpandable(true)
+						.build();
+
+		JPanel sourcePanel = borderLayoutPanel()
+						.centerComponent(tabbedPane()
+										.tabBuilder("API/Impl", apiImplPanel)
+										.mnemonic('A')
+										.add()
+										.tabBuilder("Combined", borderLayoutPanel()
+														.centerComponent(new JScrollPane(combinedTextArea))
+														.southComponent(flowLayoutPanel(FlowLayout.RIGHT)
+																		.add(button(createCopyControl(combinedTextArea))
+																						.build())
+																		.build())
+														.build())
+										.mnemonic('C')
+										.add()
+										.build())
+						.build();
+
+		JPanel sourcePackagePanel = borderLayoutPanel()
+						.northComponent(packagePanel)
+						.centerComponent(sourcePanel)
 						.build();
 
 		JSplitPane splitPane = splitPane()
 						.resizeWeight(RESIZE_WEIGHT)
 						.leftComponent(schemaTablePanel)
-						.rightComponent(textAreaPanel)
+						.rightComponent(sourcePackagePanel)
 						.build();
 
 		setLayout(borderLayout());
 		add(splitPane, BorderLayout.CENTER);
 
-		model.domainSource().addConsumer(sourceText -> {
-			sourceTextArea.setText(sourceText);
-			sourceTextArea.setCaretPosition(0);
+		model.domainApi().addConsumer(sourceText -> {
+			apiTextArea.setText(sourceText);
+			apiTextArea.setCaretPosition(0);
+		});
+		model.domainImpl().addConsumer(sourceText -> {
+			implementationTextArea.setText(sourceText);
+			implementationTextArea.setCaretPosition(0);
+		});
+		model.domainCombined().addConsumer(sourceText -> {
+			combinedTextArea.setText(sourceText);
+			combinedTextArea.setCaretPosition(0);
 		});
 		KeyEvents.builder()
 						.modifiers(InputEvent.ALT_DOWN_MASK)
@@ -171,8 +225,21 @@ public final class DomainGeneratorPanel extends JPanel {
 						.action(Control.control(definitionTable::requestFocusInWindow))
 						.enable(this)
 						.keyCode(KeyEvent.VK_3)
-						.action(Control.control(sourceTextArea::requestFocusInWindow))
+						.action(Control.control(apiTextArea::requestFocusInWindow))
+						.enable(this)
+						.keyCode(KeyEvent.VK_4)
+						.action(Control.control(implementationTextArea::requestFocusInWindow))
+						.enable(this)
+						.keyCode(KeyEvent.VK_5)
+						.action(Control.control(combinedTextArea::requestFocusInWindow))
 						.enable(this);
+	}
+
+	private static Control createCopyControl(JTextArea textArea) {
+		return Control.builder(() -> Utilities.setClipboard(textArea.getText()))
+						.name(Messages.copy())
+						.mnemonic(Messages.copy().charAt(0))
+						.build();
 	}
 
 	public void showFrame() {
@@ -196,7 +263,7 @@ public final class DomainGeneratorPanel extends JPanel {
 						.owner(this)
 						.title("Populating")
 						.northPanel(northPanel)
-						.onResult(model.schemaModel()::refresh)
+						.onResult(model.definitionModel()::refresh)
 						.execute();
 	}
 
@@ -257,13 +324,7 @@ public final class DomainGeneratorPanel extends JPanel {
 		findLookAndFeelProvider(defaultLookAndFeelName(DomainGeneratorPanel.class.getName()))
 						.ifPresent(LookAndFeelProvider::enable);
 		try {
-			Database database = Database.instance();
-			DomainGeneratorModel explorerModel = DomainGeneratorModel.domainGeneratorModel(database,
-							Dialogs.loginDialog()
-											.icon(Logos.logoTransparent())
-											.validator(user -> database.createConnection(user).close())
-											.show());
-			SwingUtilities.invokeLater(() -> new DomainGeneratorPanel(explorerModel).showFrame());
+			SwingUtilities.invokeLater(DomainGeneratorPanel::start);
 		}
 		catch (CancelException ignored) {
 			System.exit(0);
@@ -271,6 +332,21 @@ public final class DomainGeneratorPanel extends JPanel {
 		catch (Exception e) {
 			Dialogs.displayExceptionDialog(e, null);
 			System.exit(1);
+		}
+	}
+
+	private static void start() {
+		Database database = Database.instance();
+		try {
+			new DomainGeneratorPanel(DomainGeneratorModel.domainGeneratorModel(database,
+							Dialogs.loginDialog()
+											.icon(Logos.logoTransparent())
+											.validator(user -> database.createConnection(user).close())
+											.show()))
+							.showFrame();
+		}
+		catch (DatabaseException e) {
+			throw new RuntimeException(e);
 		}
 	}
 }